last update

+++ lib/features/dashboard/screens/home/views/screens/orders/models/order_model.dart
@@ -79,7 +79,7 @@ class Order {
   String? paymentStatusName;
   dynamic paymentMethodSystemName;
   String? customerCurrencyCode;
-  int? currencyRate;
+  double? currencyRate;
   dynamic orderSubTotalDiscount;
   dynamic orderShippingFees;
   dynamic paymentMethodAdditionalFees;
@@ -155,7 +155,7 @@ class Order {
     paymentStatusName = json['paymentStatusName'];
     paymentMethodSystemName = json['paymentMethodSystemName'];
     customerCurrencyCode = json['customerCurrencyCode'];
-    currencyRate = json['currencyRate'];
+    currencyRate = json['currencyRate'].toDouble();
     orderSubTotalDiscount = json['orderSubTotalDiscount'];
     orderShippingFees = json['orderShippingFees'];
     paymentMethodAdditionalFees = json['paymentMethodAdditionalFees'];
